Refactor Profile into ProfileAggregate folder
Move Profile.cs to its own ProfileAggregate folder following the
existing pattern for other aggregates. Update namespace to match
file location and update all references accordingly.
