+++ src/Commitments.Core/Model/ProfileAggregate/Profile.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Quinntyne Brown. All Rights Reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace Commitments.Core.Model;
+namespace Commitments.Core.Model.ProfileAggregate;
 
 public class Profile : BaseEntity
 {
