+++ tests/Commitments.Core.Tests/Model/ProfileTests.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Quinntyne Brown. All Rights Reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using Commitments.Core.Model;
+using Commitments.Core.Model.ProfileAggregate;
 using FluentAssertions;
 using Xunit;
